Basic/wwin-01/VinPayroll/san_pham/vposervice/service/serializers.py
from rest_framework import serializers

from .models import Claim

class ClaimSerializer(serializers.ModelSerializer):
    class Meta:
        model = Claim
        fields = '__all__'
        read_only_fields = ('created_claim','updated-claim')
