Basic/wwin-01/VinPayroll/san_pham/vposervice/requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
dataclasses==0.8
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.1.5
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
XlsxWriter==3.0.1
zipp==3.5.0
djangorestframework
markdown
django-filter
numpy